checking list empty in inspection

+++ src/main/java/uz/technocorp/ecosystem/modules/inspection/InspectionServiceImpl.java
@@ -112,7 +112,10 @@ public class InspectionServiceImpl implements InspectionService {
             paths.add(dto.orderPath());
         }
         repository.save(inspection);
-        attachmentService.deleteByPaths(paths);
+        if (!paths.isEmpty()) {
+            attachmentService.deleteByPaths(paths);
+        }
+
     }
 
     @Override
